# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Conference Management Tool API",
        "description": "API documentation for the Conference Management Tool (CMT).",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Local development server"
        }
    ],
    "paths": {
...
        "/api/v1/testimonials": {
            "get": {
                "tags": [
                    "Testimonials"
                ],
                "summary": "Get all testimonials",
                "operationId": "21bbd0a5cc6a1f1f60d40f437f1f408d",
                "responses": {
                    "200": {
                        "description": "Testimonials retrieved successfully"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Testimonials"
                ],
                "summary": "Create a testimonial",
                "operationId": "280a0d496e6d8ff5fc34fc76e57ea5f6",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "conference_id",
                                    "rating",
                                    "content"
                                ],
                                "properties": {
                                    "conference_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "rating": {
                                        "type": "integer",
                                        "example": 5,
                                        "maximum": 5,
                                        "minimum": 1
                                    },
                                    "content": {
                                        "type": "string",
                                        "example": "Excellent conference."
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Testimonial created successfully"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/testimonials/{id}": {
            "get": {
                "tags": [
                    "Testimonials"
                ],
                "summary": "Get testimonial by ID",
                "operationId": "7e1c12330c65629a7f80532686719950",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Testimonial ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Testimonial retrieved successfully"
                    },
                    "404": {
                        "description": "Testimonial not found"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Testimonials"
                ],
                "summary": "Update a testimonial",
                "operationId": "7f2ecc91796a1d75ab80e0e31116304d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Testimonial ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "rating",
                                    "content"
                                ],
                                "properties": {
                                    "rating": {
                                        "type": "integer",
                                        "example": 4,
                                        "maximum": 5,
                                        "minimum": 1
                                    },
                                    "content": {
                                        "type": "string",
                                        "example": "Very good conference."
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Testimonial updated successfully"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Testimonial not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Testimonials"
                ],
                "summary": "Delete a testimonial",
                "operationId": "f708d2c9adda09874ffb50e69182845d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Testimonial ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Testimonial deleted successfully"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Testimonial not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
...
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "Sanctum",
                "description": "Enter your Sanctum token."
            }
        }
    },
    "security": [
        {
            "sanctum": []
        }
    ]
}